+++ Source/Sitecore.FakeDb/Data/Engines/DataCommands/AddVersionCommand.cs
@@ -4,24 +4,31 @@
   using System.Linq;
   using Sitecore.Data;
   using Sitecore.Data.Items;
+  using System.Threading;
 
   public class AddVersionCommand : Sitecore.Data.Engines.DataCommands.AddVersionCommand, IDataEngineCommand
   {
-    private DataEngineCommand innerCommand = DataEngineCommand.NotInitialized;
+    private ThreadLocal<DataEngineCommand> innerCommand;
 
-    public void Initialize(DataEngineCommand command)
+    public AddVersionCommand()
     {
-      this.innerCommand = command;
+      this.innerCommand = new ThreadLocal<DataEngineCommand>();
+      this.innerCommand.Value = DataEngineCommand.NotInitialized;
+    }
+
+    public virtual void Initialize(DataEngineCommand command)
+    {
+      this.innerCommand.Value = command;
     }
 
     protected override Sitecore.Data.Engines.DataCommands.AddVersionCommand CreateInstance()
     {
-      return this.innerCommand.CreateInstance<Sitecore.Data.Engines.DataCommands.AddVersionCommand, AddVersionCommand>();
+      return this.innerCommand.Value.CreateInstance<Sitecore.Data.Engines.DataCommands.AddVersionCommand, AddVersionCommand>();
     }
 
     protected override Item DoExecute()
     {
-      var dbitem = this.innerCommand.DataStorage.GetFakeItem(this.Item.ID);
+      var dbitem = this.innerCommand.Value.DataStorage.GetFakeItem(this.Item.ID);
       var language = this.Item.Language.Name;
       var version = new Version(Item.Version.Number + 1);
 
@@ -46,7 +53,7 @@
         langValues.Add(version.Number, value);
       }
 
-      return this.innerCommand.DataStorage.GetSitecoreItem(this.Item.ID, this.Item.Language, version);
+      return this.innerCommand.Value.DataStorage.GetSitecoreItem(this.Item.ID, this.Item.Language, version);
     }
   }
 }