+++ Source/Sitecore.FakeDb/Data/Engines/DataCommands/CreateItemCommand.cs
@@ -3,27 +3,36 @@
   using Sitecore.Configuration;
   using Sitecore.Data.Items;
   using Sitecore.Diagnostics;
+  using System.Threading;
 
   public class CreateItemCommand : Sitecore.Data.Engines.DataCommands.CreateItemCommand, IDataEngineCommand
   {
-    private ItemCreator itemCreator;
+    private ThreadLocal<DataEngineCommand> innerCommand;
 
-    private DataEngineCommand innerCommand = DataEngineCommand.NotInitialized;
+    private ThreadLocal<ItemCreator> itemCreator;
+
+    public CreateItemCommand()
+    {
+      this.itemCreator = new ThreadLocal<ItemCreator>();
+
+      this.innerCommand = new ThreadLocal<DataEngineCommand>();
+      this.innerCommand.Value = DataEngineCommand.NotInitialized;
+    }
 
     public virtual void Initialize(DataEngineCommand command)
     {
-      this.innerCommand = command;
+      this.innerCommand.Value = command;
     }
 
     public ItemCreator ItemCreator
     {
-      get { return this.itemCreator ?? (this.itemCreator = new ItemCreator(this.innerCommand.DataStorage)); }
-      set { this.itemCreator = value; }
+      get { return this.itemCreator.Value ?? (this.itemCreator.Value = new ItemCreator(this.innerCommand.Value.DataStorage)); }
+      set { this.itemCreator.Value = value; }
     }
 
     protected override Sitecore.Data.Engines.DataCommands.CreateItemCommand CreateInstance()
     {
-      return this.innerCommand.CreateInstance<Sitecore.Data.Engines.DataCommands.CreateItemCommand, CreateItemCommand>();
+      return this.innerCommand.Value.CreateInstance<Sitecore.Data.Engines.DataCommands.CreateItemCommand, CreateItemCommand>();
     }
 
     protected override Item DoExecute()