+++ Source/Sitecore.FakeDb/Data/Engines/DataCommands/MoveItemCommand.cs
@@ -1,22 +1,30 @@
-﻿namespace Sitecore.FakeDb.Data.Engines.DataCommands
+﻿using System.Threading;
+namespace Sitecore.FakeDb.Data.Engines.DataCommands
 {
   public class MoveItemCommand : Sitecore.Data.Engines.DataCommands.MoveItemCommand, IDataEngineCommand
   {
-    private DataEngineCommand innerCommand = DataEngineCommand.NotInitialized;
+    private ThreadLocal<DataEngineCommand> innerCommand;
+
+    public MoveItemCommand()
+    {
+      this.innerCommand = new ThreadLocal<DataEngineCommand>();
+      this.innerCommand.Value = DataEngineCommand.NotInitialized;
+    }
 
     public virtual void Initialize(DataEngineCommand command)
     {
-      this.innerCommand = command;
+      this.innerCommand.Value = command;
     }
+
 
     protected override Sitecore.Data.Engines.DataCommands.MoveItemCommand CreateInstance()
     {
-      return this.innerCommand.CreateInstance<Sitecore.Data.Engines.DataCommands.MoveItemCommand, MoveItemCommand>();
+      return this.innerCommand.Value.CreateInstance<Sitecore.Data.Engines.DataCommands.MoveItemCommand, MoveItemCommand>();
     }
 
     protected override bool DoExecute()
     {
-      var dataStorage = this.innerCommand.DataStorage;
+      var dataStorage = this.innerCommand.Value.DataStorage;
 
       var fakeItem = dataStorage.GetFakeItem(this.Item.ID);