+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ﻿0.15.0
 * [NEW] Implemented RoleProviderSwitcher
+* [NEW] Implemented FakeAuthorizationProvider
+* [FIX] DataEngine commands became thread local to solve concurrency issue
 
 0.14.0
 * [NEW] Implemented context user switching